refactor: Abstract all `.serverless/` directory references
We want to be able to package / deploy serverless stacks in parallel.
But before we can do that we need to track down all the places that the
`.serverless/` directory is used and abstract to common functions.

The `.serverless/` directory is used as a temporary working directory,
but that prevents parallel Serverless commands from being run at the
same time. A unique temporary directory is needed for each run. In order
to maintain backwards compatibility for v2.x.x the temporary directory
needs to be separate from the package output directory.

With this change, all uses of `.serverless` either point to the package
output directory (getPackageOutputDir), or to the temporary working
directory (getTmpDir)