fix: Parallel packaging / deploying
We want to be able to package / deploy serverless stacks in parallel.

Historically the .serverless folder was used as a temporary directory, but
that prevents parallel packaging or deploying commands from being run at
the same time. A unique temporary directory is needed for each run.
In order to maintain backwards compatibility for v2.x.x the temporary
directory needs to be separate from the package output directory.

All uses of `.serverless` should either point to the package
output directory (getPackageOutputDir), or to the temp director (getTmpDir)